refactor: view 패키지 생성후 메서드 이동
