+++ src/main/kotlin/lotto/Application.kt
@@ -6,42 +6,22 @@ import lotto.util.LottoGenerator
 import lotto.util.Validator.validate1000Unit
 import lotto.util.Validator.validateInteger
 import lotto.util.Validator.validateRange
+import lotto.view.InputView
+import lotto.view.OutputView
 
 fun main() {
-    printGameStartMessage()
-    val purchaseCount = getValidateUserInput() / 1000
+    val inputView = InputView()
+    val outputView = OutputView()
+    outputView.printGameStartMessage()
+    val userInputPrice = inputView.getValidateUserInput()
+    val purchaseCount = userInputPrice / 1000
     val lottoTicket = LottoTicket(purchaseCount)
     val lottoGenerator = LottoGenerator()
     println()
-    printPurchaseCount(purchaseCount)
+    outputView.printPurchaseCount(purchaseCount)
     repeat(purchaseCount) {
         val numbers = lottoGenerator.lottoPublish()
         lottoTicket.addNumbers(numbers)
     }
-    printLottoTicket(lottoTicket)
-}
-
-fun printGameStartMessage() {
-    println("구입금액을 입력해 주세요.")
-}
-
-fun getUserInput(): String = Console.readLine()
-
-fun getValidateUserInput(): Int {
-    val userInput = getUserInput()
-    validateInteger(userInput)
-    validateRange(userInput)
-    validate1000Unit(userInput)
-    return userInput.toInt()
-}
-
-fun printPurchaseCount(purchaseCount: Int) {
-    println("${purchaseCount}개를 구매했습니다.")
-}
-
-fun printLottoTicket(lottoTicket: LottoTicket) {
-    val ticket = lottoTicket.numbers
-    repeat(ticket.size) {index ->
-        println(ticket[index].toString())
-    }
+    outputView.printLottoTicket(lottoTicket)
 }
